task: create ImagePreview component and add it to moderate page
